Create code9.py
Give one number n,write a program that checks if n is not divisible by any number from 2 to 9 

Print divisible number if the number is divisible by any of the following numbers from 2 to 9. Print indivisible number if the number n is not divisible by any of the numbers from 2 to 9.