Create code35.py
n this coding question, you need to write a program that creates a list of N Fibonacci terms. The Fibonacci sequence is an integer sequence of 0, 1, 1, 2, 3, 5, 8, and so on. The first two terms are 0 and 1, and all other terms are obtained by adding the preceding two terms.

For example, if the given N is 5, your code should print the list of five Fibonacci terms: [0, 1, 1, 2, 3].