Create code38.py
In this problem, you are given a set of numbers. Your task is to remove a list of numbers from the set if they are present in it. You will receive a single line of input containing space-separated integers. The output should be a single line containing the list of remaining numbers in ascending order


For example, if the given input is "20 40 80", your code should remove the numbers (20, 40, 80) from the set. So the output should be:

[10, 30, 50, 60, 70, 90, 100]