Create code49.py
Question: Combine Two Dictionaries
Write a program to combine two dictionaries updating values for common keys. The input will be given in four lines. The first two lines will have the keys and values of the first dictionary, and the next two lines will have the keys and values of the second dictionary. The output should be a single line containing the list of tuples of dictionary items with all the key-value pairs of two dictionaries sorted in ascending order by key.